Add files via upload
This is the code that uses FlaskForm from flask_wtf and imports various other packages from wtforms and wtforms.validators to create a new class which is used to create a form for adding new customer details to the database. The initial starter code files do not establish a connection with a database yet, but the connection with a database will be established once form validators flash error messages and everything else works without any errors.