Initial commit
This is the html file for the adding new customer page. The initial commit has HTML for displaying a form, and it extends the layout HTML file. The initial commit does not have an established connection with a database yet, but connection with database will be established once flash messages are displayed for all errors in the new customer details form and after making sure there are no other errors.