Initial commit
This is the HTML file for the message that is displayed if new customer details are successfully added to the database. In the initial commit, this is displayed if valid information has been entered into the new customer details form, rather than if the data has been successfully sent to the database. This will be updated once a connection with a database is established.