Initial commit
This CSS file will contain the main customized CSS for the web app. In the initial commit, the customized CSS only has a background fadeblue color, which is being used as the background color for the navbar. More CSS will be added once the main back-end of the program works as intended.